+++ src/Tests/Dfe.Complete.Application.Tests/QueryHandlers/Project/ListAllProjectsQueryHandlerTests.cs
@@ -6,6 +6,7 @@ using DfE.CoreLibs.Testing.AutoFixture.Customizations;
 using Dfe.Complete.Application.Projects.Interfaces;
 using Dfe.Complete.Application.Projects.Models;
 using Dfe.Complete.Application.Projects.Queries.ListAllProjects;
+using Dfe.Complete.Domain.Enums;
 using Dfe.Complete.Tests.Common.Customizations.Models;
 using MockQueryable;
 using NSubstitute.ExceptionExtensions;
@@ -103,6 +104,55 @@ namespace Dfe.Complete.Application.Tests.QueryHandlers.Project
             }
         }
 
+        [Theory]
+        [CustomAutoData(
+            typeof(OmitCircularReferenceCustomization),
+            typeof(ListAllProjectsQueryModelCustomization),
+            typeof(DateOnlyCustomization))]
+        public async Task Handle_ShouldOnlyReturnProjectsThatAreAssigned(
+            [Frozen] IListAllProjectsQueryService mockListAllProjectsQueryService,
+            ListAllProjectsQueryHandler handler,
+            IFixture fixture)
+        {
+            // Arrange
+            var totalProjects = 15;
+            var expectedProjects = 10;
+            
+            var combinedProjects = fixture.CreateMany<ListAllProjectsQueryModel>(totalProjects).ToList();
+
+            // Set first 10 as assigned, last 5 as unassigned
+            for (int i = 0; i < combinedProjects.Count; i++)
+            {
+                if (i < 10)
+                {
+                    combinedProjects[i].Project!.AssignedTo = fixture.Create<Domain.Entities.User>();
+                }
+                else
+                {
+                    combinedProjects[i].Project!.AssignedTo = null;
+                }
+            }
+
+
+            var mock = combinedProjects.BuildMock();
+
+            var query = new ListAllProjectsQuery(ProjectState.Active, ProjectType.Conversion);
+
+            mockListAllProjectsQueryService.ListAllProjects(query.ProjectStatus, query.Type)
+                .Returns(mock);
+
+            // Act
+            var result = await handler.Handle(query, default);
+
+            // Assert
+            Assert.NotNull(result);
+            Assert.True(result.IsSuccess);
+            Assert.All(result.Value!, project =>
+                Assert.False(string.IsNullOrWhiteSpace(project.AssignedToFullName)));
+            Assert.Equal(expectedProjects, result.Value!.Count);
+        }
+
+        
         [Theory]
         [CustomAutoData(
             typeof(OmitCircularReferenceCustomization),